Moving from using make to simple bash scripts; also implementing using properties
git-svn-id: file:///home/jls/openjml/repo_master/OpenJML/trunk/OpenJML@1802 1627bae4-3225-46cb-a19f-684862636fd9

+++ src/org/jmlspecs/openjml/esc/JmlEsc.java
@@ -665,10 +665,16 @@ public class JmlEsc extends JmlTreeScanner {
         try {
 
             // Pick a prover to use
-            String proverToUse = "yices";
+            String proverToUse = System.getProperty("openjml.defaultProver");
+            if (proverToUse == null) proverToUse = "yices";
             //String proverToUse = "cvc";
             //String proverToUse = "simplify";
             p = AbstractProver.getProver(context,proverToUse);
+            if (p == null) {
+                // Error is already reported
+                //log.error("esc.no.prover",proverToUse);
+                return false;
+            }
 
             if (useRetract && !p.supports().retract) { p.kill(); return true; }
             if (useCoreIds && !p.supports().unsatcore) { p.kill(); return true; }
